requirements.txt
pygame==2.5.2
pygame-vkeyboard==2.0.9
opencv-python==4.10.0.84
pillow==10.4.0
mediapipe==0.10.21